Test for DB before removing it

+++ pgesmd/test_helpers.py
@@ -129,8 +129,9 @@ class TestHelpers(unittest.TestCase):
     
     def test_database_add(self):
         """Test adding new data to the DB."""
-        os.remove(f'{PROJECT_PATH}/test/data/energy_history_test.db')
-        
+        if os.path.exists(f'{PROJECT_PATH}/test/data/energy_history_test.db'):
+            os.remove(f'{PROJECT_PATH}/test/data/energy_history_test.db')
+
         query = "SELECT value FROM hour WHERE start=?"
 
         db = EnergyHistory(path='/test/data/energy_history_test.db',
@@ -203,12 +204,10 @@ class TestHelpers(unittest.TestCase):
         hour_list = [x[0] for x in cur.fetchall()]
         week_summed_avg = int(round(sum(hour_list) / len(hour_list)))
         self.assertAlmostEqual(week_summed_avg, week_avg)
-        
 
-
-    
     def test_database_json_export(self):
-        os.remove(f'{PROJECT_PATH}/test/data/energy_history_test.db')
+        if os.path.exists(f'{PROJECT_PATH}/test/data/energy_history_test.db'):
+            os.remove(f'{PROJECT_PATH}/test/data/energy_history_test.db')
 
         db = EnergyHistory(path='/test/data/energy_history_test.db',
                            json_path='/test/data/energy_history_test.json')
